refactor: Extract LogSettings management.
