+++ packages/serverpod/lib/src/server/session.dart
@@ -8,7 +8,7 @@ import 'package:serverpod/src/server/features.dart';
 import '../cache/caches.dart';
 import '../database/database.dart';
 import '../generated/protocol.dart';
-import 'log_manager.dart';
+import 'log_manager/log_manager.dart';
 
 /// When a call is made to the [Server] a [Session] object is created. It
 /// contains all data associated with the current connection and provides
