+++ packages/serverpod/lib/src/server/websocket_request_handlers/endpoint_websocket_request_handler.dart
@@ -109,7 +109,7 @@ abstract class EndpointWebsocketRequestHandler {
             var logManager = session.serverpod.logManager;
 
             var slow = duration >=
-                logManager
+                logManager.settings
                     .getLogSettingsForStreamingSession(
                       endpoint: endpointName,
                     )
